// src/app/models/producto.model.ts
export class Producto {
  id_producto?: number;
  nombre?: string;
  resumen?: string;
  descripcion?: string;
  duracion?: string;
  imagen?: string;
  ruta_imagen?: string;
  pdf?: string;
  ruta_pdf?: string;
  /* maestro?: string; */
  observacion?: string;
  precio?: number;
  destacado?: string;
  categoria_producto_id?: number;
}